+++ changelog.md
@@ -1,3 +1,9 @@
+##### 14.07.2022 (v.0.9.1)
+
+* Sped up `learn as open` and `unlearn as open` commands.
+* Fixed delayed Shift key state when localized input is enabled.
+* Fixed web-search commands.
+
 ##### 24.12.2021 (v.0.9.0)
 * Added experimental support of internationalized command input. Non-latin characters such as спасибо are allowed when
   the current input language is not English. Set LOCALIZED_INPUT = False in the settings configuration block to disable